Limit page access of users
Based on the user designation session that will be set when the user login, chosen links on sidebar will appear.

We also added security to files. Even if the users know the file structure and try to access the page without using buttons. The pages will validate them and they are not allowed in that page, the site will log them out.